e2e/src/@module@.hub/requirements.txt
django>=1.4
django-guardian
django-userena
pil
-e git+git://github.com/toastdriven/django-tastypie.git#egg=django-tastypie